pyproject.toml
[tool.poetry]
name = "shiv-shakti-app"
version = "0.1.0"
description = "Shiv & Shakti App"
authors = ["Piyush Jain <piyush.jain@snssystem.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.4.0"
django-filter = "^24.3"
face-recognition = "^1.3.0"
dlib = "^19.24.5"
setuptools = "^72.1.0"
django-haystack = "^3.3.0"
elasticsearch = "^8.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
